+++ mopidy/mpd/protocol/music_db.py
@@ -30,6 +30,15 @@ _LIST_MAPPING = {
     'genre': 'genre',
     'performer': 'performer'}
 
+_LIST_NAME_MAPPING = {
+    'album': 'Album',
+    'albumartist': 'AlbumArtist',
+    'artist': 'Artist',
+    'composer': 'Composer',
+    'date': 'Date',
+    'genre': 'Genre',
+    'performer': 'Performer'}
+
 
 def _query_from_mpd_search_parameters(parameters, mapping):
     query = {}
@@ -246,109 +255,30 @@ def list_(context, *args):
     - does not add quotes around the field argument.
     - capitalizes the field argument.
     """
-    parameters = list(args)
-    if not parameters:
+    params = list(args)
+    if not params:
         raise exceptions.MpdArgError('incorrect arguments')
-    field = parameters.pop(0).lower()
+    field = params.pop(0).lower()
 
     if field not in _LIST_MAPPING:
         raise exceptions.MpdArgError('incorrect arguments')
 
-    if len(parameters) == 1:
+    if len(params) == 1:
         if field != 'album':
             raise exceptions.MpdArgError('should be "Album" for 3 arguments')
-        return _list_album(context, {'artist': parameters})
+        query = {'artist': params}
+    else:
+        try:
+            query = _query_from_mpd_search_parameters(params, _LIST_MAPPING)
+        except exceptions.MpdArgError as e:
+            e.message = 'not able to parse args'
+            raise
+        except ValueError:
+            return
 
-    try:
-        query = _query_from_mpd_search_parameters(parameters, _LIST_MAPPING)
-    except exceptions.MpdArgError as e:
-        e.message = 'not able to parse args'
-        raise
-    except ValueError:
-        return
-
-    if field == 'artist':
-        return _list_artist(context, query)
-    if field == 'albumartist':
-        return _list_albumartist(context, query)
-    elif field == 'album':
-        return _list_album(context, query)
-    elif field == 'composer':
-        return _list_composer(context, query)
-    elif field == 'performer':
-        return _list_performer(context, query)
-    elif field == 'date':
-        return _list_date(context, query)
-    elif field == 'genre':
-        return _list_genre(context, query)
-
-
-def _list_artist(context, query):
-    artists = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        for artist in track.artists:
-            if artist.name:
-                artists.add(('Artist', artist.name))
-    return artists
-
-
-def _list_albumartist(context, query):
-    albumartists = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        if track.album:
-            for artist in track.album.artists:
-                if artist.name:
-                    albumartists.add(('AlbumArtist', artist.name))
-    return albumartists
-
-
-def _list_album(context, query):
-    albums = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        if track.album and track.album.name:
-            albums.add(('Album', track.album.name))
-    return albums
-
-
-def _list_composer(context, query):
-    composers = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        for composer in track.composers:
-            if composer.name:
-                composers.add(('Composer', composer.name))
-    return composers
-
-
-def _list_performer(context, query):
-    performers = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        for performer in track.performers:
-            if performer.name:
-                performers.add(('Performer', performer.name))
-    return performers
-
-
-def _list_date(context, query):
-    dates = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        if track.date:
-            dates.add(('Date', track.date))
-    return dates
-
-
-def _list_genre(context, query):
-    genres = set()
-    results = context.core.library.find_exact(**query).get()
-    for track in _get_tracks(results):
-        if track.genre:
-            genres.add(('Genre', track.genre))
-    return genres
+    name = _LIST_NAME_MAPPING[field]
+    result = context.core.library.get_distinct(field, query)
+    return [(name, value) for value in result.get()]
 
 
 @protocol.commands.add('listall')