+++ docs/changelog.rst
@@ -36,6 +36,9 @@ v0.20.0 (UNRELEASED)
 - When seeking in paused state, do not change to playing state. (Fixed
   :issue:`939`)
 
+- Add :meth:`mopidy.core.LibraryController.get_distinct` for getting unique
+  values for a given field. (Fixes: :issue:`913`)
+
 **Commands**
 
 - Make the ``mopidy`` command print a friendly error message if the
@@ -97,6 +100,9 @@ v0.20.0 (UNRELEASED)
   from using ``listall`` and ``listallinfo`` which recursively lookup the entire
   "database". If you insist on using a client that needs these commands change
   :confval:`mpd/command_blacklist`.
+
+- Switch the ``list`` command over to using
+  :meth:`mopidy.core.LibraryController.get_distinct`. (Fixes: :issue:`913`)
 
 **HTTP frontend**